feat(#2683): introduce @ato-copilot/shared pnpm workspace package
De-duplicates SSE and MCP types/parsers across VS Code extension, M365
Teams bot, and Dashboard into a single source of truth.

- packages/ato-shared: SseEvent, SseEventHandler, parseSseChunk (./sse)
  and ToolExecution, ErrorDetail, McpResponse, generateConversationId (./mcp)
  with dual ESM/CJS exports
- pnpm-workspace.yaml: declares packages/* as workspace members
- extensions/vscode, extensions/m365, src/Ato.Copilot.Dashboard: remove
  local duplicate type definitions; wire @ato-copilot/shared via tsconfig
  path aliases
- M365 SSE semantic fix: event.type -> event.event, data object -> raw
  string (two JSON.parse sites + tests updated; internally consistent)
- scripts/bootstrap.sh, devcontainer.json: pnpm provisioning added
