+++ extensions/vscode/tsconfig.json
@@ -4,7 +4,6 @@
     "target": "ES2022",
     "lib": ["ES2022"],
     "outDir": "dist",
-    "rootDir": ".",
     "sourceMap": true,
     "strict": true,
     "esModuleInterop": true,
@@ -13,7 +12,11 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "declaration": true,
-    "declarationMap": true
+    "declarationMap": true,
+    "paths": {
+      "@ato-copilot/shared": ["../../packages/ato-shared/src/index.ts"],
+      "@ato-copilot/shared/*": ["../../packages/ato-shared/src/*"]
+    }
   },
   "include": ["src/**/*.ts", "test/**/*.ts"],
   "exclude": ["node_modules", "dist"]
