+++ packages/ato-shared/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@ato-copilot/shared",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Shared types and utilities for ATO Copilot TypeScript projects",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
+  "types": "./dist/esm/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js",
+      "types": "./dist/esm/index.d.ts"
+    }
+  },
+  "scripts": {
+    "build:esm": "tsc -p tsconfig.json --outDir dist/esm --module ESNext",
+    "build:cjs": "tsc -p tsconfig.json --outDir dist/cjs --module CommonJS",
+    "build": "npm run build:esm && npm run build:cjs",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^5.4.5"
+  }
+}
